+++ src/main/java/com/verdantartifice/thaumicwonders/proxy/ProxyTESR.java
@@ -1,0 +1,12 @@
+package com.verdantartifice.thaumicwonders.proxy;
+
+import com.verdantartifice.thaumicwonders.client.renderers.tile.TesrOreDiviner;
+import com.verdantartifice.thaumicwonders.common.tiles.devices.TileOreDiviner;
+
+import net.minecraftforge.fml.client.registry.ClientRegistry;
+
+public class ProxyTESR {
+    public void setupTESR() {
+        ClientRegistry.bindTileEntitySpecialRenderer(TileOreDiviner.class, new TesrOreDiviner());
+    }
+}